+++ crates/ruma-state-res/benches/state_res_bench.rs
@@ -27,7 +27,7 @@ use ruma_common::{
             join_rules::{JoinRule, RoomJoinRulesEventContent},
             member::{MembershipState, RoomMemberEventContent},
         },
-        RoomEventType, StateEventType,
+        StateEventType, TimelineEventType,
     },
     room_id, user_id, EventId, MilliSecondsSinceUnixEpoch, OwnedEventId, RoomId, RoomVersionId,
     UserId,
@@ -234,7 +234,7 @@ impl TestStore<PduEvent> {
         let create_event = to_pdu_event::<&EventId>(
             "CREATE",
             alice(),
-            RoomEventType::RoomCreate,
+            TimelineEventType::RoomCreate,
             Some(""),
             to_raw_json_value(&json!({ "creator": alice() })).unwrap(),
             &[],
@@ -246,7 +246,7 @@ impl TestStore<PduEvent> {
         let alice_mem = to_pdu_event(
             "IMA",
             alice(),
-            RoomEventType::RoomMember,
+            TimelineEventType::RoomMember,
             Some(alice().to_string().as_str()),
             member_content_join(),
             &[cre.clone()],
@@ -257,7 +257,7 @@ impl TestStore<PduEvent> {
         let join_rules = to_pdu_event(
             "IJR",
             alice(),
-            RoomEventType::RoomJoinRules,
+            TimelineEventType::RoomJoinRules,
             Some(""),
             to_raw_json_value(&RoomJoinRulesEventContent::new(JoinRule::Public)).unwrap(),
             &[cre.clone(), alice_mem.event_id().to_owned()],
@@ -270,7 +270,7 @@ impl TestStore<PduEvent> {
         let bob_mem = to_pdu_event(
             "IMB",
             bob(),
-            RoomEventType::RoomMember,
+            TimelineEventType::RoomMember,
             Some(bob().to_string().as_str()),
             member_content_join(),
             &[cre.clone(), join_rules.event_id().to_owned()],
@@ -281,7 +281,7 @@ impl TestStore<PduEvent> {
         let charlie_mem = to_pdu_event(
             "IMC",
             charlie(),
-            RoomEventType::RoomMember,
+            TimelineEventType::RoomMember,
             Some(charlie().to_string().as_str()),
             member_content_join(),
             &[cre, join_rules.event_id().to_owned()],
@@ -352,7 +352,7 @@ fn member_content_join() -> Box<RawJsonValue> {
 fn to_pdu_event<S>(
     id: &str,
     sender: &UserId,
-    ev_type: RoomEventType,
+    ev_type: TimelineEventType,
     state_key: Option<&str>,
     content: Box<RawJsonValue>,
     auth_events: &[S],
@@ -396,7 +396,7 @@ fn INITIAL_EVENTS() -> HashMap<OwnedEventId, Arc<PduEvent>> {
         to_pdu_event::<&EventId>(
             "CREATE",
             alice(),
-            RoomEventType::RoomCreate,
+            TimelineEventType::RoomCreate,
             Some(""),
             to_raw_json_value(&json!({ "creator": alice() })).unwrap(),
             &[],
@@ -405,7 +405,7 @@ fn INITIAL_EVENTS() -> HashMap<OwnedEventId, Arc<PduEvent>> {
         to_pdu_event(
             "IMA",
             alice(),
-            RoomEventType::RoomMember,
+            TimelineEventType::RoomMember,
             Some(alice().as_str()),
             member_content_join(),
             &["CREATE"],
@@ -414,7 +414,7 @@ fn INITIAL_EVENTS() -> HashMap<OwnedEventId, Arc<PduEvent>> {
         to_pdu_event(
             "IPOWER",
             alice(),
-            RoomEventType::RoomPowerLevels,
+            TimelineEventType::RoomPowerLevels,
             Some(""),
             to_raw_json_value(&json!({ "users": { alice(): 100 } })).unwrap(),
             &["CREATE", "IMA"],
@@ -423,7 +423,7 @@ fn INITIAL_EVENTS() -> HashMap<OwnedEventId, Arc<PduEvent>> {
         to_pdu_event(
             "IJR",
             alice(),
-            RoomEventType::RoomJoinRules,
+            TimelineEventType::RoomJoinRules,
             Some(""),
             to_raw_json_value(&RoomJoinRulesEventContent::new(JoinRule::Public)).unwrap(),
             &["CREATE", "IMA", "IPOWER"],
@@ -432,7 +432,7 @@ fn INITIAL_EVENTS() -> HashMap<OwnedEventId, Arc<PduEvent>> {
         to_pdu_event(
             "IMB",
             bob(),
-            RoomEventType::RoomMember,
+            TimelineEventType::RoomMember,
             Some(bob().to_string().as_str()),
             member_content_join(),
             &["CREATE", "IJR", "IPOWER"],
@@ -441,7 +441,7 @@ fn INITIAL_EVENTS() -> HashMap<OwnedEventId, Arc<PduEvent>> {
         to_pdu_event(
             "IMC",
             charlie(),
-            RoomEventType::RoomMember,
+            TimelineEventType::RoomMember,
             Some(charlie().to_string().as_str()),
             member_content_join(),
             &["CREATE", "IJR", "IPOWER"],
@@ -450,7 +450,7 @@ fn INITIAL_EVENTS() -> HashMap<OwnedEventId, Arc<PduEvent>> {
         to_pdu_event::<&EventId>(
             "START",
             charlie(),
-            RoomEventType::RoomTopic,
+            TimelineEventType::RoomTopic,
             Some(""),
             to_raw_json_value(&json!({})).unwrap(),
             &[],
@@ -459,7 +459,7 @@ fn INITIAL_EVENTS() -> HashMap<OwnedEventId, Arc<PduEvent>> {
         to_pdu_event::<&EventId>(
             "END",
             charlie(),
-            RoomEventType::RoomTopic,
+            TimelineEventType::RoomTopic,
             Some(""),
             to_raw_json_value(&json!({})).unwrap(),
             &[],
@@ -478,7 +478,7 @@ fn BAN_STATE_SET() -> HashMap<OwnedEventId, Arc<PduEvent>> {
         to_pdu_event(
             "PA",
             alice(),
-            RoomEventType::RoomPowerLevels,
+            TimelineEventType::RoomPowerLevels,
             Some(""),
             to_raw_json_value(&json!({ "users": { alice(): 100, bob(): 50 } })).unwrap(),
             &["CREATE", "IMA", "IPOWER"], // auth_events
@@ -487,7 +487,7 @@ fn BAN_STATE_SET() -> HashMap<OwnedEventId, Arc<PduEvent>> {
         to_pdu_event(
             "PB",
             alice(),
-            RoomEventType::RoomPowerLevels,
+            TimelineEventType::RoomPowerLevels,
             Some(""),
             to_raw_json_value(&json!({ "users": { alice(): 100, bob(): 50 } })).unwrap(),
             &["CREATE", "IMA", "IPOWER"],
@@ -496,7 +496,7 @@ fn BAN_STATE_SET() -> HashMap<OwnedEventId, Arc<PduEvent>> {
         to_pdu_event(
             "MB",
             alice(),
-            RoomEventType::RoomMember,
+            TimelineEventType::RoomMember,
             Some(ella().as_str()),
             member_content_ban(),
             &["CREATE", "IMA", "PB"],
@@ -505,7 +505,7 @@ fn BAN_STATE_SET() -> HashMap<OwnedEventId, Arc<PduEvent>> {
         to_pdu_event(
             "IME",
             ella(),
-            RoomEventType::RoomMember,
+            TimelineEventType::RoomMember,
             Some(ella().as_str()),
             member_content_join(),
             &["CREATE", "IJR", "PA"],
@@ -522,7 +522,7 @@ trait EventTypeExt {
     fn with_state_key(self, state_key: impl Into<String>) -> (StateEventType, String);
 }
 
-impl EventTypeExt for &RoomEventType {
+impl EventTypeExt for &TimelineEventType {
     fn with_state_key(self, state_key: impl Into<String>) -> (StateEventType, String) {
         (self.to_string().into(), state_key.into())
     }
@@ -530,7 +530,7 @@ impl EventTypeExt for &RoomEventType {
 
 mod event {
     use ruma_common::{
-        events::{pdu::Pdu, RoomEventType},
+        events::{pdu::Pdu, TimelineEventType},
         MilliSecondsSinceUnixEpoch, OwnedEventId, RoomId, UserId,
     };
     use ruma_state_res::Event;
@@ -562,7 +562,7 @@ mod event {
             }
         }
 
-        fn event_type(&self) -> &RoomEventType {
+        fn event_type(&self) -> &TimelineEventType {
             match &self.rest {
                 Pdu::RoomV1Pdu(ev) => &ev.kind,
                 Pdu::RoomV3Pdu(ev) => &ev.kind,